goodgames/requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
diff-match-patch==20121119
Django==1.11.7
django-allauth==0.34.0
django-cors-headers==2.1.0
django-extensions==1.9.7
django-import-export==0.5.1
django-rest-auth==0.9.2
django-rest-framework==0.1.0
django-rest-swagger==2.1.2
djangorestframework==3.7.3
enum34==1.1.6
et-xmlfile==1.0.1
futures==3.2.0
google-api-core==0.1.1
google-auth==1.2.1
google-cloud-core==0.28.0
google-cloud-language==1.0.0
google-cloud-storage==1.6.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpcio==1.7.0
gunicorn==19.7.1
idna==2.6
itypes==1.1.0
jdcal==1.3
Jinja2==2.10
jsonpickle==0.9.5
MarkupSafe==1.0
oauthlib==2.0.6
odfpy==1.3.5
olefile==0.44
openapi-codec==1.3.2
openpyxl==2.4.9
Pillow==4.3.0
protobuf==3.5.0.post1
psycopg2==2.7.3.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
python-openid==2.2.5
pytz==2017.3
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
simplejson==3.12.0
six==1.11.0
tablib==0.12.1
typing==3.6.2
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
wheel==0.30.0

